backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
python-socketio==5.13.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.40
pydantic==2.11.3
python-multipart==0.0.9
starlette==0.46.2
websockets==12.0
python-dotenv==1.0.1 